python/python_duplicate/requirements.txt
opencv-contrib-python
numpy
memory-profiler
Pillow
troposphere[policy]